requirements.txt
python-pptx==0.6.23
pywin32==306
openai==1.54.0
python-dotenv==1.0.0
Pillow==10.4.0
tqdm==4.66.1
tenacity==8.2.3
